Explicitly type-annotate IterableLie and TraversableLike # view
Here's the situation in the current stdlib:

In TraversableLike:

    def view = new TraversableView[A, Repr] {
      protected lazy val underlying = self.repr
      override def foreach[U](f: A => U) = self foreach f
    }

in IterableLike:

    override def view = new IterableView[A, Repr] {
      protected lazy val underlying = self.repr
      override def iterator = self.iterator
    }

In Dotty, we have to annotate the type in IterableLike because otherwise the
inherited type from TraversableLike is chosen. So the IterableLike view becomes:

    override def view: IterableView[A, Repr] = new IterableView[A, Repr] {
      protected lazy val underlying = self.repr
      override def iterator = self.iterator
    }

But now we also have to annotate the view in TraversableLike because otherwise
its inferred type is

    TraversableView[A, Repr] { val underlying: Repr }

which is not a supertype of `IterableView[A, Repr]`. Or else we add the `underlying`
refinement to `IterableLike#view`. But the intention is that `underlying` should
not be part of the type.