Minor stdlib changes needed for the tuple PR
- scala/collection/mutable/ArrayOps.scala

Because _i are added via implicit conversion, these lines now need to trigger two implicit conversions in a raw, which is not supported by scalac.

- scala/math/Ordering.scala

Somehow for the second argument of the `Tuple5` does not uses the (...) syntax. Because of the scala package it references scala.Tuple5, which means the method is not a proper override for def compare.